still trying to fix the brcm package download...

+++ prep_03_custom_debs.sh
@@ -4,13 +4,13 @@ pwd=$PWD
 
 # koreader
 cd overlays/custom_debs/
-rm koreader*.deb*
+test -e koreader-2023.08-arm64.deb && rm koreader*.deb
 wget -nv https://github.com/koreader/koreader/releases/download/v2023.08/koreader-2023.08-arm64.deb
 cd ${pwd}
 
 # evsieve
 cd overlays/custom_debs/
-rm evsieve*.deb
+test -e evsieve_1.3.1_arm64.deb && rm evsieve*.deb
 wget https://github.com/m-weigand/evsieve_pn/releases/download/v1.3.1-arm64/evsieve_1.3.1_arm64.deb
 cd "${pwd}"
 
@@ -24,14 +24,14 @@ cd "${pwd}"
 
 # pinenote dbus service
 cd overlays/custom_debs
-rm pinenote_dbus_service*.deb
+test -e pinenote_dbus_service_0.1.0_arm64.deb && rm pinenote_dbus_service*.deb
 wget -nv https://github.com/m-weigand/pinenote_dbus_service/releases/download/v20221207/pinenote_dbus_service_0.1.0_arm64.deb
 cd "${pwd}"
 
 # 2023.05.17: (temporarily) fixed brcm firmware
 cd overlays/custom_debs
 test -e firmware-brcm80211_20230625-1+salsaci_all.deb && rm firmware-brcm80211_20230625-1+salsaci_all.deb
-wget -nv https://salsa.debian.org/diederik/firmware-nonfree/-/jobs/4652912/artifacts/raw/debian/output/firmware-brcm80211_20230625-1+salsaci_all.deb?inline=false
+wget -nv -O firmware-brcm80211_20230625-1+salsaci_all.deb https://salsa.debian.org/diederik/firmware-nonfree/-/jobs/4652912/artifacts/raw/debian/output/firmware-brcm80211_20230625-1+salsaci_all.deb?inline=false
 cd "${pwd}"
 
 cd overlays/custom_debs
